Reformat into a proper cpp project
